refactor: 사건 요약 및 증인 프로필 생성을 위한 함수들을 case_generation 모듈로 분리하여 코드 구조 개선
